# WebApi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ambar WebApi",
    "version": "1.0.0",
    "description": "The Ambar Cloud WebApi"
  },
  "paths": {
...
    "/api/search/stats": {
      "get": {
        "summary": "Get documents stats by query",
        "description": "Returns statistics for documents based on the specified query.",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "description": "URI_ENCODED query string. Check details of query syntax [here](https://blog.ambar.cloud/mastering-ambar-search-queries/).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with documents statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid query."
          }
        }
      }
    },
...
    "/api/stats": {
      "get": {
        "summary": "Get Statistics",
        "description": "Retrieve statistics related to document processing, content types, and processing rates.",
        "tags": [
          "Stats"
        ],
        "responses": {
          "200": {
            "description": "Successful response with combined statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contentType": {
                      "type": "object"
                    },
                    "procRate": {
                      "type": "object"
                    },
                    "procTotal": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Failed to retrieve statistics."
          }
        }
      }
    },
    "/api/stats/combined": {
      "get": {
        "summary": "Get Combined Statistics",
        "description": "Retrieve combined statistics for files based on a query.",
        "tags": [
          "Stats"
        ],
        "responses": {
          "200": {
            "description": "Successful response with combined statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Failed to retrieve combined statistics."
          }
        }
      }
    },
    "/api/stats/processing": {
      "get": {
        "summary": "Get Processing Statistics",
        "description": "Retrieve statistics related to document processing based on a query.",
        "tags": [
          "Stats"
        ],
        "responses": {
          "200": {
            "description": "Successful response with processing statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Failed to retrieve processing statistics."
          }
        }
      }
    },
    "/api/tags": {
      "get": {
        "summary": "Get Tags",
        "description": "Retrieve a list of tags along with the count of associated files.",
        "tags": [
          "Tags"
        ],
        "responses": {
          "200": {
            "description": "Successful response with an array of tags and their associated file counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "The name of the tag."
                      },
                      "filesCount": {
                        "type": "integer",
                        "description": "The count of files associated with the tag."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Failed to retrieve tags."
          }
        }
      }
    },
    "/api/tags/{fileId}/{tagType}/{tagName}": {
      "post": {
        "summary": "Add Tag For File",
        "description": "Add a tag for a specific file.",
        "tags": [
          "Tags"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the file to add the tag to."
          },
          {
            "in": "path",
            "name": "tagType",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The type of the tag to add."
          },
          {
            "in": "path",
            "name": "tagName",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The name of the tag to add."
          }
        ],
        "requestBody": {
          "required": false
        },
        "responses": {
          "200": {
            "description": "Successful response with updated tags.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Required field is missing or invalid tag type."
          },
          "500": {
            "description": "Internal Server Error. Failed to add tag."
          }
        }
      },
      "delete": {
        "summary": "Delete Tag From File",
        "description": "Delete a tag from a specific file.",
        "tags": [
          "Tags"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the file to delete the tag from."
          },
          {
            "in": "path",
            "name": "tagType",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The type of the tag to delete."
          },
          {
            "in": "path",
            "name": "tagName",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The name of the tag to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with updated tags.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Required field is missing."
          },
          "500": {
            "description": "Internal Server Error. Failed to delete tag."
          }
        }
      }
    },
...
  },
...
}